Refactor collection creator help out to simplify.
